Improved compatibility for entrypoint.sh

+++ entrypoint.sh
@@ -18,17 +18,19 @@ fi
 nv_prefix=/app/.venv/lib/python3.12/site-packages/nvidia
 
 workdir=$(pwd)
-if [ ! -L "${nv_prefix}/cudnn/lib/libcudnn.so" ]; then
+if [ -d "${nv_prefix}/cudnn/lib" ] && [ ! -L "${nv_prefix}/cudnn/lib/libcudnn.so" ]; then
     cd "${nv_prefix}/cudnn/lib"; \
     ln -s libcudnn.so.9 libcudnn.so; \
 fi
-if [ ! -L "${nv_prefix}/cublas/lib/libcublas.so" ]; then
+if [ -d "${nv_prefix}/cublas/lib" ] && [ ! -L "${nv_prefix}/cublas/lib/libcublas.so" ]; then
     cd "${nv_prefix}/cublas/lib"; \
     ln -s libcublas.so.12 libcublas.so;
 fi
 cd $workdir
 
-export LD_LIBRARY_PATH=${nv_prefix}/cudnn/lib:${nv_prefix}/cublas/lib
+if [ -d "${nv_prefix}/cudnn/lib" ] || [ -d "${nv_prefix}/cublas/lib" ]; then
+    export LD_LIBRARY_PATH=${nv_prefix}/cudnn/lib:${nv_prefix}/cublas/lib
+fi
 
 # Ensure target correct
 if [ "prompt" = "${1}" ]; then
